# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JetKöy API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List available jobs in the pool
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of available jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new job in the pool
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}/grab:
    post:
      operationId: grabJob
      tags: [jobs]
      summary: Grab a job from the pool
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job grabbed successfully with contact details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrabJobResponse"
        "400":
          description: Insufficient credits or job already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/my:
    get:
      operationId: getMyJobs
      tags: [jobs]
      summary: Get jobs created by current user and jobs grabbed by current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: My jobs summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyJobsSummary"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Job:
      type: object
      properties:
        id:
          type: string
        creatorId:
          type: string
        creatorName:
          type: string
        creatorPlate:
          type: string
        departure:
          type: string
        destination:
          type: string
        passengerName:
          type: string
        passengerPhoneMasked:
          type: string
        totalFare:
          type: number
        commission:
          type: number
        status:
          type: string
          enum: [available, grabbed]
        createdAt:
          type: string
      required:
        - id
        - creatorId
        - creatorName
        - creatorPlate
        - departure
        - destination
        - passengerName
        - passengerPhoneMasked
        - totalFare
        - commission
        - status
        - createdAt
    JobDetail:
      type: object
      properties:
        id:
          type: string
        creatorId:
          type: string
        creatorName:
          type: string
        creatorPlate:
          type: string
        departure:
          type: string
        destination:
          type: string
        passengerName:
          type: string
        passengerPhone:
          type: string
        totalFare:
          type: number
        commission:
          type: number
        status:
          type: string
          enum: [available, grabbed]
        grabbedById:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - creatorId
        - creatorName
        - creatorPlate
        - departure
        - destination
        - passengerName
        - passengerPhone
        - totalFare
        - commission
        - status
        - createdAt
    GrabJobResponse:
      type: object
      properties:
        job:
          $ref: "#/components/schemas/JobDetail"
        passengerPhone:
          type: string
        newCreditBalance:
          type: number
      required:
        - job
        - passengerPhone
        - newCreditBalance
    CreateJobRequest:
      type: object
      properties:
        departure:
          type: string
        destination:
          type: string
        passengerName:
          type: string
        passengerPhone:
          type: string
        totalFare:
          type: number
        commission:
          type: number
      required:
        - departure
        - destination
        - passengerName
        - passengerPhone
        - totalFare
        - commission
    MyJobsSummary:
      type: object
      properties:
        createdJobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        grabbedJobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
      required:
        - createdJobs
        - grabbedJobs
...